requirements.txt
aniso8601==7.0.0
Click==7.0
Flask==1.1.1
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pika==1.1.0
pytz==2019.2
requests==2.22.0
scipy-1.3.1
six==1.12.0
sounddevice==0.3.13
Werkzeug==0.15.5
